pyproject.toml
[tool.poetry]
name = "cinema_app"
version = "0.1.0"
description = ""
authors = ["mandeep kumar <mandeepkumar1652@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.3"
django-ninja = "^1.1.0"
celery = "^5.3.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.1"
pillow = "^10.2.0"
django-redis = "^5.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
